refactor: separate printer methods
