+++ docs/README.md
@@ -15,8 +15,9 @@
 - [X] input 에러 핸들링: 출발역&도착역 존재 여부
 - [X] input 에러 핸들링: 출발역&도착역 같은 경우
 - [X] input 에러 핸들링: 출발역&도착역 연결 여부  
-- [ ] input 에러 핸들링: 기타
+- [ ] ~input 에러 핸들링: 기타~
 - [X] 최단거리 경로 구하기
 - [X] 최소시간 경로 구하기
 - [X] 결과 출력 
 - [X] (added) 상수화
+- [X] (added) 테이블 출력 함수 분리
