+++ src/index.js
@@ -1,14 +1,16 @@
 import { stations } from './data.js';
-import { makeDijkstra, printResult, validateInput, alertNoRoute } from './utils/helpers.js';
+import { makeDijkstra, validateInput, alertNoRoute } from './utils/helpers.js';
+import { printResult } from './utils/print.js';
 import { SEARCH_TYPE } from './constants.js';
 
 export default function SubwayPath() {
+  this.stations = stations;
+
   const dijkstraDist = makeDijkstra(SEARCH_TYPE.DIST, stations);
   const dijkstraTime = makeDijkstra(SEARCH_TYPE.TIME, stations);
   const departStation = document.getElementById('departure-station-name-input');
   const arriveStation = document.getElementById('arrival-station-name-input');
   const submitBtn = document.getElementById('search-button');
-
   submitBtn.addEventListener('click', () => showResult());
 
   const showResult = () => {
